+++ photutils/morphology/tests/test_core.py
@@ -31,7 +31,7 @@ def test_data_properties():
     props = data_properties(data, mask=None)
     props2 = data_properties(data, mask=mask)
     properties = ['xcentroid', 'ycentroid', 'area']
-    result = [props[i].value for i in properties]
-    result2 = [props2[i].value for i in properties]
+    result = [getattr(props, i).value for i in properties]
+    result2 = [getattr(props2, i).value for i in properties]
     assert_allclose([0.5, 0.5, 4.0], result, rtol=0, atol=1.e-6)
     assert_allclose([0.5, 0.0, 2.0], result2, rtol=0, atol=1.e-6)